Updated samples
