+++ ojdbc-provider-jackson-oson/pom.xml
@@ -17,8 +17,11 @@
     <dependencies>
         <dependency>
             <groupId>com.oracle.database.jdbc</groupId>
+            <artifactId>ojdbc-provider-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.oracle.database.jdbc</groupId>
             <artifactId>ojdbc8</artifactId>
-            <version>23.6.0.24.09.240830</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
