+++ ojdbc-provider-samples/pom.xml
@@ -14,6 +14,11 @@
     <version>1.0.1</version>
   </parent>
 
+  <properties>
+    <maven.compiler.target>11</maven.compiler.target>
+    <maven.compiler.source>11</maven.compiler.source>
+  </properties>
+
   <dependencies>
     <dependency>
       <groupId>com.oracle.database.jdbc</groupId>
@@ -30,7 +35,7 @@
       <artifactId>ojdbc-provider-gcp</artifactId>
       <version>${project.parent.version}</version>
     </dependency>
-        <dependency>
+    <dependency>
       <groupId>com.oracle.database.jdbc</groupId>
       <artifactId>ojdbc-provider-jackson-oson</artifactId>
       <version>${project.parent.version}</version>
