+++ ojdbc-provider-samples/src/main/java/oracle/jdbc/provider/oson/sample/model/Emp.java
@@ -1,4 +1,4 @@
-package oracle.jdbc.provider.oson.model;
+package oracle.jdbc.provider.oson.sample.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
